[REFACTOR] Move Description component as a shared component between stages
